Saving only best epoch results, use stitch new
