+++ main.py
@@ -345,9 +345,11 @@ def runTraining(args):
                             warnings.filterwarnings('ignore', category=UserWarning)
                             predicted_class: Tensor = probs2class(pred_probs)
                             mult: int = 63 if K == 5 else (255 / (K - 1))
+                            # Decide per-epoch save dir (temporary if save_best_only)
+                            base_iter_dir = (args.dest / "_tmp" / f"iter{e:03d}") if args.save_best_only else (args.dest / f"iter{e:03d}")
                             save_images(predicted_class * mult,
                                         data['stems'],
-                                        args.dest / f"iter{e:03d}" / m)
+                                        base_iter_dir / m)
 
                     j += B  # Keep in mind that _in theory_, each batch might have a different size
                     # For the DSC average: do not take the background class (0) into account:
@@ -401,7 +403,10 @@ def runTraining(args):
             best_folder = args.dest / "best_epoch"
             if best_folder.exists():
                 rmtree(best_folder)
-            copytree(args.dest / f"iter{e:03d}", Path(best_folder))
+            # Source depends on whether we saved only to temporary folder
+            src_iter_dir = (args.dest / "_tmp" / f"iter{e:03d}") if args.save_best_only else (args.dest / f"iter{e:03d}")
+            if src_iter_dir.exists():
+                copytree(src_iter_dir, Path(best_folder))
 
             torch.save(net, args.dest / "bestmodel.pkl")
             torch.save(net.state_dict(), args.dest / "bestweights.pt")
@@ -409,6 +414,16 @@ def runTraining(args):
             # Save model artifacts to wandb
             wandb.save(str(args.dest / "bestmodel.pkl"))
             wandb.save(str(args.dest / "bestweights.pt"))
+        
+        # Cleanup temporary per-epoch images if not best and saving best only
+        if args.save_best_only:
+            tmp_dir = args.dest / "_tmp" / f"iter{e:03d}"
+            if tmp_dir.exists() and (current_dice <= best_dice):
+                try:
+                    rmtree(tmp_dir)
+                except Exception:
+                    print(f">> Warning: Failed to remove temporary directory {tmp_dir}")
+                    pass
     
     # Final logging and artifact saving
     print(f">>> Training completed. Best dice: {best_dice:.3f}")
@@ -464,6 +479,9 @@ def main():
                         help="Custom name for wandb experiment run")
     parser.add_argument('--seed', type=int, default=42,
                         help="Random seed for reproducibility (default: 42)")
+    
+    parser.add_argument('--save_best_only', action='store_true',
+                        help="If set, only store images for best_epoch; intermediate epoch images are temporary")
     parser.add_argument('--alter_enet', action='store_true',
                     help="Apply paper-faithful ENet tweaks (bias-free convs, SpatialDropout2d, BN+PReLU on initial conv, final fullconv)")
     parser.add_argument('--attn', type=str, default=None,
